Explicitly set the bitness of the jni lib required dep
JNI libraries which are not embedded inside android_app are installed by
adding the name of the library to `LOCAL_REQUIRED_MODULES` of the app.
The bitness resolution is currently missing in
build/make/core/soong_app_prebuilt.mk, due to which both variants (if
available) gets installed. The second variant is likely unusable
anyways, since the .apk does not contain a symlink to the .so file on
/system/lib/<libfoo>.so

Suffix the bitness of the dependency explicilty so that that extraneous
variants do not get installed.

Test: go build ./java
Test: presubmits
Bug: 369678122
Change-Id: I91c1359c6060e4617a2ad1ca50b551ccfe25ee64