Add install_symlink_host Soong module type
The `install_symlink_host` module type allows Soong to create symbolic
links on the host side. This functionality addresses the issue where a
`cc_binary` with `symlink_preferred_arch: true` cannot be used as a
tool in a `cc_genrule`.
The solution involves setting `symlink_preferred_arch` to `false` for
the `tool` and then using `install_symlink_host` to explicitly create
the originally intended symbolic link.

Bug: 342330305
Test: Unit test and CI
Change-Id: I850060be25bdea3aa37febaf499a30c18cfd181b