Add 'dsts' property to prebuilt_etc
The current set of properties for prebuilts aren't quite flexible
enough to properly install some types of prebuilts. If we've got a
prebuilt where some of the files need to be installed to one directory
and some to another directory then it's impossible to specify.
Similarly while it's possible to change the name of the installed file
when dealing with just one file, it's impossible to change the name
when dealing with more than one file.

Fix this by adding the 'dsts' property which allows you to specify
where each source file should end up.

Bug: 361834699
Test: m nothing --no-skip-soong-tests
Test: Build some actual firmware that uses this new feature
Change-Id: If35d9520790a4e540193cb9570ce4014e20f8049