Add install_symlink_host Soong module type
The `install_symlink_host` module type allows Soong to create symbolic
links on the host side. This functionality addresses the issue where a
`cc_binary` with `symlink_preferred_arch: true` cannot be used as a
tool in a `cc_genrule`.
The solution involves setting `symlink_preferred_arch` to `false` for
the `tool` and then using `install_symlink_host` to explicitly create
the originally intended symbolic link.

Bug: 342330305
Test: 1. Add install_symlink_host in Android.bp
      2. m the module and check in stall to $(HOST_OUT)/
      or Uint test
Change-Id: Ib7aa8ec462e0aed8e9196a519be592206469dbcd