rust: Don't produce apex variants if apex_exclude
The apex variants of apex_exclude modules were leading to a mixing of
apex and non-apex dependencies in their dependency trees, causing
StableId collisions when an -L path for both variants was emitted
to the rustc command.

Since we already exclude these variants from apexes (they are
effectively serving as a stop-gap for stub support), don't produce apex
variants if apex_exclude is true.

Bug: 203478530
Test: m
Change-Id: I1448ce1aad989b8a97fd31552a00afa7e6b78829