Skip resource verification on intermediate android libraries
The use case for this is kythe builds, which still use the legacy
resource processor. https://r.android.com/3207856 special-cased kythe
builds to use the legacy resource processor and retain R.srcjar xrefs.
This unfortunately broke xrefs for android apps that have android
library dependencies with unresolved resource references.

With this CL, the resource verification will be done only on the
top-level app for both ResourceProcessorBusyBox and the legacy
processor builds.

Test: XREF_CORPUS=blah m Traceur-res # builds now
Test: presubmits
Bug: 354854007
Change-Id: Ibc0eaa220a2eb233078c325318ef3bece243f375