Do not install internal files of apex
The motivation for this is to prevent the internal deps of apexes from
being installed in the top-level `android_filesystem`.

This CL does not get rid of the internal androidmk modules created by
apexes. Symbols listed in `LOCAL_SOONG_JNI_LIBS_SYMBOLS` will be copied
to $PRODUCT_OUT/apex/<apex> if <apex> is built. However, these
additional files will not be registed in ctx.packagingSpecs of the apex.

Test: go test ./apex
Test: no changes in the merged `jacoco-report-classes-all.jar` of
coverage builds

Change-Id: I5ca4b8069af6c4b464fd13e9a6553379eef3a57b