Modify disabled flags to be exposed in keep-flagged-apis
Currently, keep-flagged-apis filters flags that are disabled (so that
the apis associated with the disabled flags are not exposed in the
"exportable" stubs), but sdk_with_runtime_apis product builds require
disabled flagged apis to be exposed.

Filtering of the flags should be done by aconfig, and keep-flagged-apis
should only convert the list of flags to the metalava-consumable format.

Test: lunch sdk_with_runtime_apis-trunk_staging-eng && m sdk dist and observe outputs
Bug: 371248797
Change-Id: If1622a6b81bc69a99e8f89c60a11ebec7899b95b