Add cc_preprocess_no_configuration modules
This is intended to replace the cc_object modules in external/cronet,
which are only used for the c preprocessor. However, they have java
modules that consume their result, so they have a dependency from the
common arch to the first arch. This only worked because of the 1-variant
fallback, which we're trying to remove, so add an archless module that
runs clang that they can use.

Bug: 372091092
Test: go test
Change-Id: I4fff3eabb7e31d2888b6e70c39310113acf5590e