Switch from genrule to java_genrule
These genrules had java dependencies. java modules have os/arch
variants whereas regular genrules do not. The dependency worked because
of the "1-variant fallback" feature of song that we're trying to
remove to support incremental analysis. Switch to java_genrule so that
the variants of the genrule and its dependencies match.

Bug: 372091092
Test: m nothing
Change-Id: I928c1939db59ee59994cc1ed3a7410e0b2e1dd7a